Remove noacpi numa=noacpi and noefi from c2 kernel options.
These cause issues with the xhci driver. With these options the
kernel hangs when accessing the hokuyo and FTDI USB-serial adapter
connected to c2. Without these options I can't see anything wrong
in the running kernel. (and I've tested many kernels, with and
without lowlatency to eliminate other possible causes of this
issue.)

Signed-off-by: Matthieu Herrb <matthieu.herrb@laas.fr>