add custom umountfs to work around system freeze on shutdown
As send out by Clearpath, the second server, c2, failed to shutdown
and froze during the shutdown procedure.

The problematic point turned out to be the unmounting of the local filesystems.
As specified in debian's umountfs script, all required mount points are unmounted
simultanously. However, on c2's configuration, that stacks nfs and unionfs mounts,
this triggers the freeze. This could be a kernel bug, but at the moment, there
is no way to test this with a current kernel...

I changed it to unmount the mount points in a for-loop and now c2 turns off
/ reboots as expected. Nothing else changed and if this is not required anymore,
this diversion can be removed again.